Update code for SS4
