+++ composer.json
@@ -1,20 +1,20 @@
 {
-  "name": "silverstripe/multiuser-editing-alert",
-  "type": "silverstripe-module",
-  "description": "A module that indicates when people are editing the same page in the CMS",
-  "license": "BSD-3-Clause",
-  "keywords": ["silverstripe", "multi-user", "concurrent", "editing", "alert", "user", "content", "author"],
-  "authors": [
-    {
-      "name": "Julian Seidenberg",
-      "homepage": "http://www.deltaflow.com"
+    "name": "silverstripe/multiuser-editing-alert",
+    "type": "silverstripe-module",
+    "description": "A module that indicates when people are editing the same page in the CMS",
+    "license": "BSD-3-Clause",
+    "keywords": ["silverstripe", "multi-user", "concurrent", "editing", "alert", "user", "content", "author"],
+    "authors": [
+        {
+            "name": "Julian Seidenberg",
+            "homepage": "http://www.deltaflow.com"
+        }
+    ],
+    "require": {
+        "silverstripe/cms": "^4.0@dev"
+    },
+    "minimum-stability": "dev",
+    "extra": {
+        "installer-name": "multiuser-editing"
     }
-  ],
-  "require": {
-    "silverstripe/cms": ">=3.1"
-  },
-  "minimum-stability": "dev",
-  "extra": {
-    "installer-name": "multiuser-editing"
-  }
 }
