+++ css/multiuserediting.css
@@ -1,144 +1,161 @@
+.breadcrumbs-wrapper {
+    overflow: visible;
+    width: 176px;
+}
+
 /* list of all users separated by <br> instead of newline */
 .multi-user-list {
-	list-style: inherit;
-	margin-left: 25px;
-	margin-top: 10px;
+    list-style: inherit;
+    margin-left: 0;
+    margin-top: 10px;
 }
 
 /* Multiuser CMS tree icon placement */
 .multi-user-icon {
-	float: right;
-	position: relative;
-	height: 22px; /* Set height for proper icon placement of child element (img) */
+    float: right;
+    position: relative;
+    height: 22px; /* Set height for proper icon placement of child element (img) */
+    width: 22px;
 }
 
-	/* Center verticaly icon */
-	.multi-user-icon > img {
-		position: absolute;
-		top: 50%;
-		-webkit-transform: translate(0px,-50%);
-		transform: translate(0px,-50%);
-	}
+/* Center verticaly icon */
+.multi-user-icon > img {
+    position: absolute;
+    top: 50%;
+    -webkit-transform: translate(0px,-50%);
+    transform: translate(0px,-50%);
+}
 
 
 /* Float a CMS header title to left for proper placement
 of '.multi-user-message-wrap' */
 .cms-content-header h2 {
-	display: inline-block;
-	margin: 0;
+    display: inline-block;
+    margin: 0;
 }
 
 /* Division wrapper for multiediting icon,
 placement: on right side of ".breadcrumbs-wrapper" */
 .multi-user-message-wrap {
-	float: right;
-	height: 40px; /* Match height of CMS header for proper message placement */
-	margin-left: 15px;
-	margin-right: 0;
-	padding-top: 6px; /* Center icon in vertical of CMS header */
-	position: relative;
-	width: 40px; /* Equal width to height for better  */
+    float: right;
+    height: 40px; /* Match height of CMS header for proper message placement */
+    margin-left: 15px;
+    margin-right: 0;
+    padding-top: 6px; /* Center icon in vertical of CMS header */
+    position: relative;
+    width: 40px; /* Equal width to height for better  */
+    z-index: 3;
 }
 
-	/* Hide message box by default */
-	.multi-user-editing-alert-message {
-		position: absolute;
-		top: 100%;
-		left: -9999px;
-		width: 340px;
-		margin: 0;
-		opacity: 0;
-		-ms-filter: "progid:DXImageTransform.Microsoft.Alpha(Opacity=0)";
-		-webkit-transition: opacity 0.3s ease-in-out 0s, left 0 0.3s;
-		transition: opacity 0.3s ease-in-out 0s, left 0 0.3s;
-	}
+/* Hide message box by default */
+.multi-user-editing-alert-message {
+    position: absolute;
+    right: -9999px;
+    top: 100%;
+    width: 270px;
+    margin: 0;
+    opacity: 0;
+    -ms-filter: "progid:DXImageTransform.Microsoft.Alpha(Opacity=0)";
+    -webkit-transition: opacity 0.3s ease-in-out 0s, left 0 0.3s;
+    transition: opacity 0.3s ease-in-out 0s, left 0 0.3s;
+}
 
-	/* Show message box on hover/focus of icon */
-	.multi-user-message-wrap:hover .multi-user-editing-alert-message,
-	.multi-user-message-wrap:focus .multi-user-editing-alert-message {
-		left: 0;
-		opacity: 1;
-		-ms-filter: "progid:DXImageTransform.Microsoft.Alpha(Opacity=100)";
-		-webkit-transition: opacity 0.3s ease-in-out 0s;
-		transition: opacity 0.3s ease-in-out 0s;
-	}
+@media (min-width: 768px) {
+    .multi-user-editing-alert-message {
+        width: 340px;
+    }
+}
 
-		/* Add a top caret pointer for message box */
-		.multi-user-editing-top-caret {
-			position: absolute;
-			content: "";
-			left: 4px;
-			top: -10px;
-			width: 0;
-			height: 0;
-			border-style: solid;
-			border-width: 0 7px 9px 7px;
-			border-color: transparent transparent #e68288 transparent;
-		}
+/* Show message box on hover/focus of icon */
+.multi-user-message-wrap:hover .multi-user-editing-alert-message,
+.multi-user-message-wrap:focus .multi-user-editing-alert-message {
+    opacity: 1;
+    right: 0;
+    -ms-filter: "progid:DXImageTransform.Microsoft.Alpha(Opacity=100)";
+    -webkit-transition: opacity 0.3s ease-in-out 0s;
+    transition: opacity 0.3s ease-in-out 0s;
+}
 
-		.multi-user-editing-top-caret:before {
-			border-color: rgba(0, 0, 0, 0) rgba(0, 0, 0, 0) #fae8e9;
-			border-style: solid;
-			border-width: 0 6px 8px;
-			content: "";
-			height: 0;
-			left: -6px;
-			position: absolute;
-			top: 2px;
-			width: 0;
-		}
+/* Add a top caret pointer for message box */
+.multi-user-editing-top-caret {
+    position: absolute;
+    content: "";
+    right: 20px;
+    top: -10px;
+    width: 0;
+    height: 0;
+    border-style: solid;
+    border-width: 0 7px 9px 7px;
+    border-color: transparent transparent #e68288 transparent;
+}
 
-			/* Styling of common box text */
-			.multi-user-editing-label {
-				font-size: 14px;
-				line-height: 1.5;
-			}
+.multi-user-editing-top-caret:before {
+    border-color: rgba(0, 0, 0, 0) rgba(0, 0, 0, 0) #fae8e9;
+    border-style: solid;
+    border-width: 0 6px 8px;
+    content: "";
+    height: 0;
+    left: -6px;
+    position: absolute;
+    top: 2px;
+    width: 0;
+}
+
+/* Styling of common box text */
+.multi-user-editing-label {
+    font-size: 14px;
+    line-height: 1.5;
+}
+
+.multi-user-editing-alert-message .multi-user-list li {
+    display: list-item;
+}
 
 /* OPTIONAL shake animation for CMS header icon,
 Usage: add a class="animated shake" on <img> on line 99 of multiUserEditing.js;
 Atribution: https://github.com/daneden/animate.css */
 .animated {
-	-webkit-animation-duration: 1.2s;
-	animation-duration: 1.2s;
-	-webkit-animation-fill-mode: both;
-	animation-fill-mode: both;
+    -webkit-animation-duration: 1.2s;
+    animation-duration: 1.2s;
+    -webkit-animation-fill-mode: both;
+    animation-fill-mode: both;
 }
 
 @-webkit-keyframes shake {
-	from, to {
-		-webkit-transform: translate3d(0, 0, 0);
-		transform: translate3d(0, 0, 0);
-	}
+    from, to {
+        -webkit-transform: translate3d(0, 0, 0);
+        transform: translate3d(0, 0, 0);
+    }
 
-	10%, 30%, 50%, 70%, 90% {
-		-webkit-transform: translate3d(-5px, 0, 0);
-		transform: translate3d(-5px, 0, 0);
-	}
+    10%, 30%, 50%, 70%, 90% {
+        -webkit-transform: translate3d(-5px, 0, 0);
+        transform: translate3d(-5px, 0, 0);
+    }
 
-	20%, 40%, 60%, 80% {
-		-webkit-transform: translate3d(5px, 0, 0);
-		transform: translate3d(5px, 0, 0);
-	}
+    20%, 40%, 60%, 80% {
+        -webkit-transform: translate3d(5px, 0, 0);
+        transform: translate3d(5px, 0, 0);
+    }
 }
 
 @keyframes shake {
-	from, to {
-		-webkit-transform: translate3d(0, 0, 0);
-		transform: translate3d(0, 0, 0);
-	}
+    from, to {
+        -webkit-transform: translate3d(0, 0, 0);
+        transform: translate3d(0, 0, 0);
+    }
 
-	10%, 30%, 50%, 70%, 90% {
-		-webkit-transform: translate3d(-5px, 0, 0);
-		transform: translate3d(-5px, 0, 0);
-	}
+    10%, 30%, 50%, 70%, 90% {
+        -webkit-transform: translate3d(-5px, 0, 0);
+        transform: translate3d(-5px, 0, 0);
+    }
 
-	20%, 40%, 60%, 80% {
-		-webkit-transform: translate3d(5px, 0, 0);
-		transform: translate3d(5px, 0, 0);
-	}
+    20%, 40%, 60%, 80% {
+        -webkit-transform: translate3d(5px, 0, 0);
+        transform: translate3d(5px, 0, 0);
+    }
 }
 
 .shake {
-	-webkit-animation-name: shake;
-	animation-name: shake;
+    -webkit-animation-name: shake;
+    animation-name: shake;
 }
