+++ src/MultiUserEditingController.php
@@ -1,41 +1,51 @@
-<?php 
+<?php
+
+namespace MultiUserEditing;
+
+use SilverStripe\Control\Controller;
+use SilverStripe\Core\Config\Config;
+use SilverStripe\Core\Flushable;
+use SilverStripe\Core\Injector\Injector;
+use SilverStripe\Security\Security;
+use Psr\SimpleCache\CacheInterface;
 
 class MultiUserEditingController extends Controller implements Flushable
 {
-
-    public static $allowed_actions = array(
+    private static $allowed_actions = [
         'set',
         'get'
-    );
-    
+    ];
+
     private $user = null; //current user DataObject
     protected $usersEditing = null; //array of all users editing
-    protected $editingCache = null; //SS_Cache to store all concurrently editing users
+    protected $editingCache = null; //CacheInterface to store all concurrently editing users
 
-    public function init()
+    protected function init()
     {
         parent::init();
 
         //check user login status
-        $this->user = Member::currentUser();
+        $this->user = Security::getCurrentUser();
         if (!$this->user) {
             user_error("User needs to be logged in access multi-user editing data", E_USER_ERROR);
         }
 
         //get the cache data
-        $this->editingCache = SS_Cache::factory('multiuserediting');
-        $usersEditing = unserialize($this->editingCache->load('editing'));
+        $this->editingCache = Injector::inst()->get(CacheInterface::class . '.multiuserediting');
+
+
+        $usersEditing = unserialize($this->editingCache->get('editing'));
 
         //create a new simple PHP object to store user editing data
         if(!$usersEditing) {
             $usersEditing = array();
         }
-        
+
         $timeout = Config::inst()->get(get_class($this), 'userTimeoutInSeconds');
-        
+
         //remove any users that have timed out
         foreach($usersEditing as $id => $user) {
-            if (!empty($user['lastEdited']) && 
+            if (!empty($user['lastEdited']) &&
                 strtotime($user['lastEdited']) < strtotime('-'.$timeout.' seconds')) {
                 //user has timed out after above number of minutes
                 unset($usersEditing[$id]);
@@ -47,17 +57,18 @@ class MultiUserEditingController extends Controller implements Flushable
 
     public static function flush()
     {
-        SS_Cache::factory('multiuserediting')->clean(Zend_Cache::CLEANING_MODE_ALL);
+        $cache = Injector::inst()->get(CacheInterface::class . '.multiuserediting');
+        $cache->delete('multiuserediting');
     }
 
     public function set($request)
     {
         $pageID = $request->param('ID');
-        
+
         if (!intval($pageID)) {
             user_error("No page ID provided", E_USER_ERROR);
         }
-        
+
         $dataArray = array();
         $dataArray['lastEdited'] = date("Y-m-d H:i:s"); //when this was last updated
         $dataArray['abbreviatedName'] = $this->user->FirstName . ' ' . substr($this->user->Surname, 0, 1);
@@ -68,13 +79,13 @@ class MultiUserEditingController extends Controller implements Flushable
 
         //update the user editing data structure for the current user
         $this->usersEditing[$this->user->ID] = $dataArray;
-        
+
         //save to the cache
-        $this->editingCache->save(serialize($this->usersEditing), 'editing');
-        
+        $this->editingCache->set('editing', serialize($this->usersEditing));
+
         return $this->get();
     }
-    
+
     public function get()
     {
         $refresh = array(
